fix(storybook): align the stories with the app, and stop overselling the guard
Three-way review of the annotation PR found the convention certifying
things it had not checked.

The onboarding stories mounted the hero inside `maxWidth: 720; padding:
48px 32px` — a frame the app never renders. The hero actually lands in
ChatView's empty slot wrapped in `.maka-onboarding-surface`, and that
class plus two `:has(.maka-firstrun)` rules own its height, padding and
alignment. None of those cascaded in the story, so its geometry was its
own scaffolding — which is precisely how #1433 produced a 135px offset
and a centring bug that neither reproduced in the built app. The story
now uses the app's class chain; measured in Storybook, both cascade
rules apply for the first time.

`CommandPaletteEmpty` claimed "⌘K in a context that offers no commands".
No such context exists — `buildCommandList` always emits 新建对话 first,
and the palette's empty branch says so in its own copy ("没有匹配的命令 /
换个关键词"). It is now `CommandPaletteNoMatch`, entered through a query,
which is the state a user can actually reach. That one annotation was
written, reviewed and merged as confidently as the other 123, and the
contract passed it — so the contract now says what it does:

  story-reachable-path-contract → story-annotation-contract

with a header that leads on the limit ("it cannot check the sentence is
TRUE — only a reviewer following the call chain can"). Two escape
hatches close with it: `satisfies Story` exports were invisible to the
export scan, and a fixture object with a `title` field above `meta`
would have exempted a whole file. The third test asserted
`length >= 0`, which no input can fail; it now demonstrates the
exemption is load-bearing instead. Repo-root lookup moves to the one
`findRepoRoot` the neighbouring contract already had, replacing a fixed
`../../../../..` chain that resolves to `/` if the build output moves.

`app-shell.stories.tsx` told readers the hero appears whenever
`sessions.length === 0`. Removing the ready_empty panel narrowed that
gate to unfinished setup, so a configured user with no sessions now
lands on EmptyHome — the sibling PR invalidated an annotation the
annotation PR never touched.

The convention itself moves to apps/desktop/stories/FIDELITY.md. It had
shipped as a ~10-line block pasted into every product story file and
drifted on day one: 17 byte-identical copies plus an older wording in
app-shell.stories.tsx, with the contract only grepping for the literal
string "FIDELITY CONVENTION". Each file now points at the doc, and the
doc says what the copies did not — the frame matters, not just the
component.

Refs #1433
