+++ packages/ui/stories/skills-panel.stories.tsx
@@ -2,15 +2,8 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 import { SkillsModuleMain } from '../src/skills-panel.js';
 import type { SkillEntry } from '../src/module-panel-types.js';
 
-// FIDELITY CONVENTION (#1433) — every story in this file must map to an app
-// state a real user can reach, with that path noted above the story. Stories
-// are treated as ground truth for what the product looks like, so one that
-// composes an unreachable state makes every visual comparison built on it
-// wrong. If the app changes and a story no longer matches a reachable state,
-// fix the story or delete it — do not keep both "the app" and "the story
-// version" of a surface alive. Where a story deliberately puts several states
-// side by side for review, say so: the arrangement is a scaffold, each panel
-// is the reachable state.
+// Fidelity convention (#1433): every story below names the real app path
+// that reaches it. See apps/desktop/stories/FIDELITY.md.
 
 const meta = {
   title: 'Product/Skills',
